text refactored
